rename project to neat-html
