--- docs/formatting_html.md
@@ -1,61 +1,0 @@
-### Block
-```html
-<p>
-</p>
-
-<p>
-    Hello
-</p>
-```
-
-### Inline
-```html
-<span></span>
-
-<span>Hello</span>
-```
-
-### Block > Block
-```html
-<div>
-    <div>
-        Hello
-    </div>
-</div>
-
-<div>
-    Hello
-    <div>
-        Hello
-    </div>
-</div>
-```
-
-### Inline > Inline
-```html
-<span><span>Hello</span></span>
-
-<span>Hello<span>Hello</span></span>
-```
-
-### Block > Inline
-```html
-<div>
-    <span>Hello</span>
-</div>
-```
-
-### Inline > Block
-```html
-<span><div>
-    Hello
-</div></span>
-```
-
-### Block > Block, Block
-```html
-<div>
-    <div></div>
-    <div></div>
-</div>
-```
